+++ R/LearnerS4KNN.R
@@ -66,7 +66,7 @@ setMethod("is_available", signature(object="familiarKNN"),
 
 #####get_default_hyperparameters,KNNlinear#####
 setMethod("get_default_hyperparameters", signature(object="familiarKNN"),
-          function(object, data=NULL){
+          function(object, data=NULL, ...){
             
             # Initialise list and declare hyperparameter entries.
             param <- list()
@@ -169,7 +169,8 @@ setMethod("..train", signature(object="familiarKNN", data="dataObject"),
                                           data=data@data,
                                           k=object@hyperparameters$k,
                                           method=as.character(object@hyperparameters$distance_metric),
-                                          scale=FALSE))
+                                          scale=FALSE),
+                              error=identity)
             
             # Check if the model trained at all.
             if(inherits(model, "error")) return(callNextMethod())